Refactor assets and files into separate commands
