+++ src/mloq/assets/docs/makefile_docs.txt
@@ -7,6 +7,7 @@ SPHINXOPTS    ?=
 SPHINXBUILD   ?= sphinx-build
 SOURCEDIR     = source
 BUILDDIR      = build
+NOTEBOOKS_DIR = .
 
 # Put it first so that "make" without argument is like "make help".
 help:
