+++ src/mloq/assets/project/main.txt
