+++ src/mloq/assets/project/version.txt
