+++ src/mloq/assets/shared/Makefile
@@ -27,21 +27,6 @@ test:
 test-codecov:
 	find -name "*.pyc" -delete
 	pytest -n $n -s -o log_cli=true -o log_cli_level=info --cov=./ --cov-report=xml --cov-config=pyproject.toml{% endif %}
-{% if package.pipenv %}
-.PHONY: pipenv-install
-pipenv-install:
-	rm -rf *.egg-info && rm -rf build && rm -rf __pycache__
-	rm -f Pipfile && rm -f Pipfile.lock
-	pipenv install --dev -r requirements-test.txt
-	pipenv install --pre --dev -r requirements-lint.txt
-	pipenv install -r requirements.txt
-	pipenv install -e .
-	pipenv lock
-
-.PHONY: pipenv-test
-pipenv-test:
-	find -name "*.pyc" -delete
-	pipenv run pytest -s{% endif %}
 {% if docker.makefile %}
 .PHONY: docker-shell
 docker-shell:
