--- src/mloq/assets/templates/WHAT_MLOQ_GENERATED.md
@@ -1,4 +1,0 @@
-# What mloq generated for your project
-
-{% for file, description in generated_files %}* `{{ file }}` - {{ description }}
-{% endfor %}
