+++ src/mloq/commands/license.py
@@ -2,24 +2,46 @@
 from pathlib import Path
 
 import click
-from omegaconf import DictConfig, MISSING
+from omegaconf import DictConfig
 
 from mloq.command import Command
 from mloq.config.param_patch import param
-from mloq.files import code_of_conduct, contributing, dco, LICENSES
+from mloq.files import ASSETS_PATH, file
 
 
-OPEN_SOURCE_FILES = [dco, contributing, code_of_conduct]
+TEMPLATES_PATH = ASSETS_PATH / "license"
+dco = file(
+    "DCO.md",
+    TEMPLATES_PATH,
+    "Developer Certificate of Origin - needed in open source projects to certify that "
+    "the incoming contributions are legitimate",
+    is_static=True,
+)
+mit_license = file("MIT_LICENSE", TEMPLATES_PATH, "license of the project", dst="LICENSE")
+apache_license = file("APACHE_LICENSE", TEMPLATES_PATH, "license of the project", dst="LICENSE")
+gpl_license = file(
+    "GPL_LICENSE",
+    TEMPLATES_PATH,
+    "license of the project",
+    dst="LICENSE",
+    is_static=True,
+)
+LICENSES = {
+    "MIT": mit_license,
+    "Apache-2.0": apache_license,
+    "GPL-3.0": gpl_license,
+}
+
+LICENSE_FILES = [file for file in LICENSES.values()] + [dco]
 
 
 class LicenseCMD(Command):
     """Implement the functionality of the license Command."""
 
     cmd_name = "license"
-    files = tuple([file for file in LICENSES.values()] + OPEN_SOURCE_FILES)
+    files = tuple(LICENSE_FILES)
     LICENSES = LICENSES
     disable = param.Boolean(default=None, doc="Disable license command?")
-    open_source = param.Boolean(MISSING, doc="Is the project Open Source?")
     license = param.String("MIT", doc="Project license type")
     copyright_year = param.Integer("${current_year:}", doc="Year when the project started")
     copyright_holder = param.String("${globals.owner}", doc="Copyright holder")
@@ -35,8 +57,7 @@ class LicenseCMD(Command):
     def record_files(self) -> None:
         """Register the files that will be generated by mloq."""
         conf = self.record.config.license
-        if conf.open_source:
-            for file in OPEN_SOURCE_FILES:
-                self.record.register_file(file=file, path=Path())
+        self.record.register_file(file=dco, path=Path())
+        if conf.license != "proprietary":
             license_file = self.LICENSES[conf.license]
             self.record.register_file(file=license_file, path=Path())
