+++ src/mloq/commands/lint.py
@@ -6,17 +6,23 @@ from omegaconf import DictConfig
 
 from mloq.command import Command
 from mloq.config.param_patch import param
-from mloq.files import lint_req, pyproject_toml
+from mloq.files import ASSETS_PATH, file, pyproject_toml
 
 
-PACKAGE_FILES = [pyproject_toml, lint_req]
+lint_req = file(
+    "requirements-lint.txt",
+    ASSETS_PATH / "lint",
+    "list of exact versions of the packages used to check your code style",
+    is_static=True,
+)
+LINT_FILES = [pyproject_toml, lint_req]
 
 
 class LintCMD(Command):
     """Implement the functionality of the lint Command."""
 
     cmd_name = "lint"
-    files = tuple(PACKAGE_FILES)
+    files = tuple(LINT_FILES)
     disable = param.Boolean(default=None, doc="Disable lint command?")
     docstring_checks = param.Boolean(True, doc="Apply docstring checks?")
     pyproject_extra = param.String("", doc="Additional pyproject.toml configuration")
@@ -30,5 +36,5 @@ class LintCMD(Command):
 
     def record_files(self) -> None:
         """Register the files that will be generated by mloq."""
-        for file in self.files:
-            self.record.register_file(file=file, path=Path())
+        for _file in self.files:
+            self.record.register_file(file=_file, path=Path())
