+++ src/mloq/commands/package.py
@@ -5,11 +5,19 @@ import click
 from omegaconf import DictConfig, MISSING, OmegaConf
 
 from mloq.command import Command
-from mloq.commands.ci import PYTHON_VERSIONS
 from mloq.config.param_patch import param
-from mloq.files import pyproject_toml, setup_py
+from mloq.files import ASSETS_PATH, file, pyproject_toml
 
 
+PACKAGE_ASSETS_PATH = ASSETS_PATH / "package"
+PYTHON_VERSIONS = ["3.6", "3.7", "3.8", "3.9", "3.10"]
+DEFAULT_PYTHON_VERSIONS = ["3.7", "3.8", "3.9", "3.10"]
+setup_py = file(
+    "setup.txt",
+    PACKAGE_ASSETS_PATH,
+    "Python package installation metadata",
+    dst="setup.py",
+)
 PACKAGE_FILES = [pyproject_toml, setup_py]
 
 
@@ -37,7 +45,7 @@ class PackageCMD(Command):
     author = param.String(doc="Author(s) of the project")
     email = param.String(doc="Owner contact email")
     python_versions = param.ListSelector(
-        default=PYTHON_VERSIONS,
+        default=DEFAULT_PYTHON_VERSIONS,
         doc="Supported python versions",
         objects=PYTHON_VERSIONS,
     )
@@ -59,5 +67,5 @@ class PackageCMD(Command):
 
     def record_files(self) -> None:
         """Register the files that will be generated by mloq."""
-        for file in self.files:
-            self.record.register_file(file=file, path=Path())
+        for _file in self.files:
+            self.record.register_file(file=_file, path=Path())
