+++ src/mloq/commands/project.py
@@ -6,19 +6,72 @@ from omegaconf import MISSING
 
 from mloq.command import Command
 from mloq.config.param_patch import param
-from mloq.files import (
-    codecov,
-    gitignore,
-    init,
-    main,
-    makefile,
-    pre_commit_hook,
-    readme,
-    test_main,
-    test_req,
-    version,
-)
+from mloq.files import ASSETS_PATH, file, makefile
 
+
+PROJECT_ASSETS_PATH = ASSETS_PATH / "project"
+readme = file("README.md", PROJECT_ASSETS_PATH, "README")
+gitignore = file(
+    ".gitignore",
+    PROJECT_ASSETS_PATH,
+    "list of files and directories ignored by Git operations",
+    is_static=True,
+)
+pre_commit_hook = file(
+    ".pre-commit-config.yaml",
+    PROJECT_ASSETS_PATH,
+    "Git pre-commit hooks configuration",
+    is_static=True,
+)
+init = file(
+    "init.txt",
+    PROJECT_ASSETS_PATH,
+    "Python package header",
+    dst="__init__.py",
+    is_static=True,
+)
+main = file(
+    "main.txt",
+    PROJECT_ASSETS_PATH,
+    "Python package executable entry point",
+    dst="__main__.py",
+    is_static=True,
+)
+test_main = file(
+    "test_main.txt",
+    PROJECT_ASSETS_PATH,
+    "Unit test of the python package executable entry point",
+    dst="test_main.py",
+    is_static=False,
+)
+test_req = file(
+    "requirements-test.txt",
+    PROJECT_ASSETS_PATH,
+    "list of exact versions of the packages needed to run your test suite",
+    is_static=True,
+)
+version = file(
+    "version.txt",
+    PROJECT_ASSETS_PATH,
+    "defines the version of the package that is incremented on each push",
+    dst="version.py",
+    is_static=True,
+)
+code_of_conduct = file(
+    "CODE_OF_CONDUCT.md",
+    PROJECT_ASSETS_PATH,
+    "behavioral rules and norms in open source projects",
+)
+codecov = file(
+    ".codecov.yml",
+    PROJECT_ASSETS_PATH,
+    "configuration of CodeCov service to track the code coverage",
+)
+contributing = file(
+    "CONTRIBUTING.md",
+    PROJECT_ASSETS_PATH,
+    "technical manual on how to contrib to the open source project",
+)
 
 PROJECT_FILES = [
     codecov,
@@ -31,6 +84,7 @@ PROJECT_FILES = [
     version,
     test_req,
     pre_commit_hook,
+    contributing,
 ]
 
 
@@ -52,6 +106,7 @@ class ProjectCMD(Command):
     def directories(self) -> Tuple[Path]:
         """Tuple containing Paths objects representing the directories created by the command."""
         project_folder = self.record.config.project.project_name.replace(" ", "_")
+
         return tuple([Path("src") / project_folder, Path("tests")])
 
     def record_files(self) -> None:
@@ -67,5 +122,5 @@ class ProjectCMD(Command):
         self.record.register_file(file=init, path=Path("tests"), description=description)
         self.record.register_file(file=test_main, path=Path("tests"))
         root_files = [readme, makefile, test_req, pre_commit_hook, codecov, gitignore]
-        for file in root_files:
-            self.record.register_file(file=file, path=Path())
+        for _file in root_files:
+            self.record.register_file(file=_file, path=Path())
