+++ tests/examples/mloq.yaml
@@ -13,7 +13,6 @@ license:
   license: MIT
   copyright_year: 2020
   copyright_holder: ${globals.owner}
-  open_source: ${globals.open_source}
   project_name: ${globals.project_name}
   email: ${globals.email}
   project_url: ${globals.project_url}
@@ -38,6 +37,8 @@ docs:
   author: ${globals.author}
   copyright_holder: ${license.copyright_holder}
   copyright_year: ${license.copyright_year}
+  default_branch: ${globals.default_branch}
+  deploy_docs: true
 
 git:
   disable: true
